+++ tosho_amap/src/lib.rs
@@ -77,7 +77,7 @@ impl AMClient {
         );
         headers.insert(
             reqwest::header::HOST,
-            reqwest::header::HeaderValue::from_static(&API_HOST),
+            reqwest::header::HeaderValue::from_static(API_HOST),
         );
         let constants = get_constants(1);
         headers.insert(
@@ -132,7 +132,7 @@ impl AMClient {
     where
         T: serde::de::DeserializeOwned + std::clone::Clone,
     {
-        let endpoint = format!("{}{}", &*BASE_API, endpoint);
+        let endpoint = format!("{}{}", BASE_API, endpoint);
 
         let mut cloned_json = json.clone().unwrap_or_default();
         self.apply_json_object(&mut cloned_json)?;
@@ -401,7 +401,7 @@ impl AMClient {
         let mut headers = make_header(&self.config, self.constants)?;
         headers.insert(
             "Host",
-            reqwest::header::HeaderValue::from_static(&IMAGE_HOST),
+            reqwest::header::HeaderValue::from_static(IMAGE_HOST),
         );
         headers.insert(
             "User-Agent",
@@ -443,7 +443,7 @@ impl AMClient {
         );
         headers.insert(
             reqwest::header::HOST,
-            reqwest::header::HeaderValue::from_static(&API_HOST),
+            reqwest::header::HeaderValue::from_static(API_HOST),
         );
         let constants = get_constants(1);
         headers.insert(
@@ -478,7 +478,7 @@ impl AMClient {
         let req = session
             .request(
                 reqwest::Method::POST,
-                format!("{}/iap/remainder.json", &*BASE_API),
+                format!("{}/iap/remainder.json", BASE_API),
             )
             .headers(make_header(&temp_config, android_c)?)
             .json(&json_body)
@@ -513,7 +513,7 @@ impl AMClient {
         let req = session
             .request(
                 reqwest::Method::POST,
-                format!("{}/{}", &*BASE_API, &*MASKED_LOGIN),
+                format!("{}/{}", BASE_API, MASKED_LOGIN),
             )
             .headers(make_header(&temp_config, android_c)?)
             .json(&json_body_login)
@@ -544,7 +544,7 @@ impl AMClient {
         let req = session
             .request(
                 reqwest::Method::POST,
-                format!("{}/iap/remainder.json", &*BASE_API),
+                format!("{}/iap/remainder.json", BASE_API),
             )
             .headers(make_header(&temp_config, android_c)?)
             .json(&json_body_session)
@@ -600,21 +600,21 @@ fn make_header(
     let mut req_headers = reqwest::header::HeaderMap::new();
 
     let current_unix = chrono::Utc::now().timestamp();
-    let av = format!("{}/{}", &*APP_NAME, constants.version);
+    let av = format!("{}/{}", APP_NAME, constants.version);
     let formulae = format!("{}{}{}", config.token(), current_unix, av);
 
     let formulae_hashed = <Sha256 as Digest>::digest(formulae.as_bytes());
     let formulae_hashed = format!("{formulae_hashed:x}");
 
     req_headers.insert(
-        HEADER_NAMES.s.as_str(),
+        HEADER_NAMES.s,
         formulae_hashed
             .parse()
             .map_err(|e| make_error!("Failed to parse custom hash into header value: {}", e))?,
     );
     if !config.identifier().is_empty() {
         req_headers.insert(
-            HEADER_NAMES.i.as_str(),
+            HEADER_NAMES.i,
             config
                 .identifier()
                 .parse()
@@ -622,7 +622,7 @@ fn make_header(
         );
     }
     req_headers.insert(
-        HEADER_NAMES.n.as_str(),
+        HEADER_NAMES.n,
         current_unix.to_string().parse().map_err(|e| {
             make_error!(
                 "Failed to parse current unix timestamp into header value: {}",
@@ -631,7 +631,7 @@ fn make_header(
         })?,
     );
     req_headers.insert(
-        HEADER_NAMES.t.as_str(),
+        HEADER_NAMES.t,
         config
             .token()
             .parse()